# GG-30386 [IGNITE-13293] .NET: Fix enum serialization performance