--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/BenchmarkEnum.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/BenchmarkEnum.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Copyright 2019 GridGain Systems, Inc. and Contributors.
  *
  * Licensed under the GridGain Community Edition License (the "License");
@@ -14,22 +14,15 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.BenchmarkDotNet
+namespace Apache.Ignite.BenchmarkDotNet.Models
 {
-    using Apache.Ignite.BenchmarkDotNet.ThinClient;
-    using global::BenchmarkDotNet.Running;
-
     /// <summary>
-    /// Benchmark runner.
+    /// Benchmark model: custom enum.
     /// </summary>
-    public static class Program
+    public enum BenchmarkEnum
     {
-        /// <summary>
-        /// Main.
-        /// </summary>
-        public static void Main()
-        {
-            BenchmarkRunner.Run<ThinClientCachePutBenchmark>();
-        }
+        Foo,
+        Bar,
+        Baz
     }
 }