--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithEnumField.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithEnumField.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Copyright 2019 GridGain Systems, Inc. and Contributors.
  *
  * Licensed under the GridGain Community Edition License (the "License");
@@ -14,22 +14,13 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.BenchmarkDotNet
+namespace Apache.Ignite.BenchmarkDotNet.Models
 {
-    using Apache.Ignite.BenchmarkDotNet.ThinClient;
-    using global::BenchmarkDotNet.Running;
-
     /// <summary>
-    /// Benchmark runner.
+    /// Benchmark model: class with a custom enum field.
     /// </summary>
-    public static class Program
+    public class ClassWithEnumField
     {
-        /// <summary>
-        /// Main.
-        /// </summary>
-        public static void Main()
-        {
-            BenchmarkRunner.Run<ThinClientCachePutBenchmark>();
-        }
+        public BenchmarkEnum BenchmarkEnum { get; set; } 
     }
 }