--- a/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithIntField.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Models/ClassWithIntField.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Copyright 2019 GridGain Systems, Inc. and Contributors.
  *
  * Licensed under the GridGain Community Edition License (the "License");
@@ -14,22 +14,14 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.BenchmarkDotNet
+// ReSharper disable UnusedAutoPropertyAccessor.Global
+namespace Apache.Ignite.BenchmarkDotNet.Models
 {
-    using Apache.Ignite.BenchmarkDotNet.ThinClient;
-    using global::BenchmarkDotNet.Running;
-
     /// <summary>
-    /// Benchmark runner.
+    /// Benchmark model: class with an int field.
     /// </summary>
-    public static class Program
+    public class ClassWithIntField
     {
-        /// <summary>
-        /// Main.
-        /// </summary>
-        public static void Main()
-        {
-            BenchmarkRunner.Run<ThinClientCachePutBenchmark>();
-        }
+        public int Foo { get; set; }
     }
 }