--- a/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryWriter.cs
@@ -903,8 +903,15 @@ namespace Apache.Ignite.Core.Impl.Binary
             _stream.WriteInt(desc.TypeId);
             _stream.WriteInt(val);
 
-            var metaHnd = _marsh.GetBinaryTypeHandler(desc);
-            SaveMetadata(desc, metaHnd.OnObjectWriteFinished());
+            var binaryTypeHolder = Marshaller.GetCachedBinaryTypeHolder(desc.TypeId);
+            if (binaryTypeHolder == null || !binaryTypeHolder.IsSaved)
+            {
+                // Save enum fields only once - they can't change locally at runtime.
+                var metaHnd = _marsh.GetBinaryTypeHandler(desc);
+                var binaryFields = metaHnd.OnObjectWriteFinished();
+                
+                SaveMetadata(desc, binaryFields);
+            }
         }
 
         /// <summary>