* Don't send enum metadata more than once (main part of this fix - see `BinaryWriter` changes)
* Cache name-value map per enum type
* Add benchmark

As a result, enum field serialization is on par with int field:

```
|                Method |     Mean |    Error |   StdDev | Ratio | RatioSD |
|---------------------- |---------:|---------:|---------:|------:|--------:|
|          PutPrimitive | 45.47 us | 0.907 us | 1.330 us |  1.00 |    0.00 |
|     PutPrimitiveAsync | 66.28 us | 1.317 us | 1.617 us |  1.45 |    0.05 |
|  PutClassWithIntField | 68.34 us | 1.275 us | 1.130 us |  1.49 |    0.05 |
| PutClassWithEnumField | 71.13 us | 1.419 us | 2.734 us |  1.57 |    0.08 |
```